refactor checklists
