+++ editor.planx.uk/src/@planx/components/Checklist/Public/components/ChecklistItems.tsx
@@ -6,7 +6,7 @@ import React from "react";
 import FormWrapper from "ui/public/FormWrapper";
 import ChecklistItem from "ui/shared/ChecklistItem/ChecklistItem";
 
-import { ChecklistLayout } from "./VisibleChecklist";
+import { ChecklistLayout } from "../Public";
 
 interface Props {
   nonExclusiveOptions: Option[];
@@ -50,7 +50,7 @@ export const ChecklistItems = ({
             checkbox
           />
         </Grid>
-      )
+      ),
     )}
   </>
 );
