+++ editor.planx.uk/src/@planx/components/Checklist/Public/tests/Public.test.tsx
@@ -4,8 +4,7 @@ import { setup } from "testUtils";
 import { vi } from "vitest";
 import { axe } from "vitest-axe";
 
-import { ChecklistLayout } from "../components/VisibleChecklist";
-import Checklist from "../Public";
+import Checklist, { ChecklistLayout } from "../Public";
 import { groupedOptions, options } from "./mockOptions";
 import { pressContinue, pressOption } from "./testUtils";
 
@@ -20,7 +19,7 @@ describe("Checklist Component - Grouped Layout", () => {
         text="home type?"
         handleSubmit={handleSubmit}
         groupedOptions={groupedOptions}
-      />
+      />,
     );
 
     await user.click(screen.getByText("Section 1"));
@@ -44,7 +43,7 @@ describe("Checklist Component - Grouped Layout", () => {
         handleSubmit={handleSubmit}
         previouslySubmittedData={{ answers: ["S1_Option1", "S3_Option1"] }}
         groupedOptions={groupedOptions}
-      />
+      />,
     );
 
     expect(screen.getByTestId("group-0-expanded")).toBeTruthy();
@@ -65,7 +64,7 @@ describe("Checklist Component - Grouped Layout", () => {
         description=""
         text="home type?"
         groupedOptions={groupedOptions}
-      />
+      />,
     );
     const results = await axe(container);
     expect(results).toHaveNoViolations();
@@ -81,7 +80,7 @@ describe("Checklist Component - Grouped Layout", () => {
         text="home type?"
         handleSubmit={handleSubmit}
         groupedOptions={groupedOptions}
-      />
+      />,
     );
     const [section1Button, section2Button, section3Button] =
       screen.getAllByRole("button");
@@ -138,7 +137,7 @@ describe("Checklist Component - Basic & Images Layout", () => {
           text="home type?"
           handleSubmit={handleSubmit}
           options={options[type]}
-        />
+        />,
       );
 
       expect(screen.getByRole("heading")).toHaveTextContent("home type?");
@@ -166,7 +165,7 @@ describe("Checklist Component - Basic & Images Layout", () => {
           handleSubmit={handleSubmit}
           previouslySubmittedData={{ answers: ["flat_id", "house_id"] }}
           options={options[type]}
-        />
+        />,
       );
 
       await user.click(screen.getByTestId("continue-button"));
@@ -183,7 +182,7 @@ describe("Checklist Component - Basic & Images Layout", () => {
           description=""
           text="home type?"
           options={options[type]}
-        />
+        />,
       );
       const results = await axe(container);
       expect(results).toHaveNoViolations();
@@ -199,7 +198,7 @@ describe("Checklist Component - Basic & Images Layout", () => {
           text="home type?"
           handleSubmit={handleSubmit}
           options={options[type]}
-        />
+        />,
       );
 
       await user.tab();
